rag-chatbot-backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
sentence-transformers==2.2.2
httpx==0.25.0
python-dotenv==1.0.0
pypdf2==3.0.1
pdfplumber==0.10.3
python-multipart==0.0.6